admin: factor out partition routes
